Improve error reporting
Augeas provides details about errors through the aug_error functions; to
use them we create a new error struct and fill that with details.

To make all that a bit more usable, we roll std::ffi::NulError into that
error struct with error code NulString. This requires a change to the
augeas_sys bindings